mgmt: Change value types
Value type for the CO2, IAQ as well and light intensity were changed to integer

+++ management-server/src/sensors/si1145/visible.ts
@@ -32,7 +32,7 @@ const VIS_INTENSITY_PARAM_NAME = "Intensity";
 /** Visible light intensity parameter measurement unit */
 const VIS_INTENSITY_PARAM_UNIT = "lux";
 /** Visible light intensity parameter value type */
-const VIS_INTENSITY_PARAM_TYPE = NumberType.FLOAT;
+const VIS_INTENSITY_PARAM_TYPE = NumberType.INTEGER;
 /** Visible light intensity parameter quality levels table */
 const VIS_INTENSITY_PARAM_QUALITY_TABLE: QualityTable = [];
